fix a bug where the document size didn't equal the bounds
A common source for this occurred when legacy scrollers were enabled.
The bounds for the collection view would remain the same, but the
usable area of the scroll view (the document visible rect) would
change. To fix this, we now check the document visible rect’s size
instead of the bounds when determining if the collection view needs a
redraw.