Nil guards for index paths passed to primitive cell (de-)selection methods
Calling selectItemAtIndexPath:animated: with indexPath = nil results in a NSInvalidArgumentException when it tries to add the index path to self.selectedIndexes. This currently happens when the collection view's data source returns an empty list of items and allowsEmptySelection is off, in which case the index path returned by indexPathForNextSelectableItemAfterIndexPath: is nil and then passed to selectItemAtIndexPath:animated: in reloadData.