Do not pass up mouse actions on cell
Until now the cell forwarded mouse down/up events that it handled to
the super implementation. This means that the events are forwarded to
the parent views underneath the cell.

This patch removes the super calls. The cell is handling the calls and
should not forward the actions to the view underneath.

This allows to differentiate between clicks on a cell and clicks in
non-cell space.